vulnhawk-dashboard-3/requirements.txt
requests==2.31.0
python-nmap==0.7.1
dnspython==2.6.1 